requested changes from review
