+++ app/inbox/[notificationId]/page.tsx
@@ -1,21 +1,10 @@
-"use client";
-
 import Link from "next/link";
-import { useEffect, useState } from "react";
-import { useParams } from "next/navigation";
+import { redirect } from "next/navigation";
+import { headers } from "next/headers";
+import { auth } from "@/lib/auth";
 import BogIcon from "@/components/bog/BogIcon/BogIcon";
 import NotificationTag from "@/components/notifications/NotificationTag";
-import api from "@/lib/api";
-
-interface NotificationDetails {
-  id: string;
-  userId: string;
-  organizationId: string;
-  createdAt: string;
-  read: boolean;
-  type: string;
-  text: string;
-}
+import { NotificationService } from "@/lib/services/NotificationService";
 
 function formatDate(dateString: string) {
   const date = new Date(dateString);
@@ -29,81 +18,28 @@ function formatDate(dateString: string) {
   });
 }
 
-export default function NotificationDetailPage() {
-  const params = useParams<{ notificationId: string }>();
-  const notificationId = params.notificationId;
+interface NotificationDetailPageProps {
+  params: Promise<{ notificationId: string }>;
+}
 
-  const [notification, setNotification] = useState<NotificationDetails | null>(
-    null,
-  );
-  const [isLoading, setIsLoading] = useState(true);
-  const [errorMessage, setErrorMessage] = useState<string | null>(null);
+export default async function NotificationDetailPage({
+  params,
+}: NotificationDetailPageProps) {
+  const session = await auth.api.getSession({ headers: await headers() });
+  if (!session?.user) redirect("/login");
 
-  useEffect(() => {
-    let isActive = true;
+  const organizationId = session.session.activeOrganizationId;
+  if (!organizationId) redirect("/");
 
-    async function loadNotification() {
-      try {
-        const response = await api.notifications[":id"].$get({
-          param: { id: notificationId },
-        });
+  const { notificationId } = await params;
+  const notification = await NotificationService.findById(notificationId);
 
-        if (!response.ok) {
-          throw new Error("Could not load notification");
-        }
-
-        const json = await response.json();
-
-        if (!isActive) {
-          return;
-        }
-
-        setNotification(json);
-        setErrorMessage(null);
-      } catch {
-        if (!isActive) {
-          return;
-        }
-
-        setNotification(null);
-        setErrorMessage("Could not load this notification.");
-      } finally {
-        if (isActive) {
-          setIsLoading(false);
-        }
-      }
-    }
-
-    void loadNotification();
-
-    return () => {
-      isActive = false;
-    };
-  }, [notificationId]);
-
-  if (isLoading) {
-    return (
-      <div className="mx-auto w-full max-w-6xl px-4 py-8 sm:px-6 lg:px-8 animate-pulse">
-        <div className="mb-6 h-5 w-44 rounded bg-grey-fill-weak" />
-        <div className="rounded-xl border border-grey-stroke-weak p-8">
-          <div className="mb-6 flex items-start justify-between gap-4">
-            <div className="h-8 w-36 rounded-full bg-grey-fill-weak" />
-            <div className="h-5 w-52 rounded bg-grey-fill-weak" />
-          </div>
-          <div className="space-y-3">
-            <div className="h-5 w-full rounded bg-grey-fill-weak" />
-            <div className="h-5 w-full rounded bg-grey-fill-weak" />
-            <div className="h-5 w-3/4 rounded bg-grey-fill-weak" />
-          </div>
-        </div>
-      </div>
-    );
-  }
-
-  if (errorMessage) {
+  if (!notification) {
     return (
       <div className="mx-auto flex w-full max-w-6xl flex-col gap-4 px-4 py-8 sm:px-6 lg:px-8">
-        <p className="text-paragraph-2 text-grey-text-weak">{errorMessage}</p>
+        <p className="text-paragraph-2 text-grey-text-weak">
+          Could not load this notification.
+        </p>
         <Link
           href="/inbox"
           className="inline-flex items-center gap-2 text-paragraph-2 text-grey-text-strong hover:text-brand-text"
@@ -115,8 +51,16 @@ export default function NotificationDetailPage() {
     );
   }
 
-  if (!notification) {
-    return null;
+  if (
+    notification.userId !== session.user.id ||
+    notification.organizationId !== organizationId
+  ) {
+    redirect("/inbox");
+  }
+
+  // Mark as read if it isn't already
+  if (!notification.read) {
+    await NotificationService.updateReadStatus(notificationId, true);
   }
 
   return (
@@ -133,7 +77,7 @@ export default function NotificationDetailPage() {
         <div className="mb-6 flex items-start justify-between gap-4">
           <NotificationTag type={notification.type} variant="light" />
           <span className="text-paragraph-2 text-grey-text-weak">
-            {formatDate(notification.createdAt)}
+            {formatDate(notification.createdAt.toString())}
           </span>
         </div>
